Yohan

+++ src/main/java/mg/itu/vokye/service/TypeEmployeService.java
@@ -22,11 +22,11 @@ public class TypeEmployeService {
         return repository.findAll();
     }
 
-    public TypeEmploye getTypeEmployeById(Long id) {
+    public TypeEmploye getTypeEmployeById(int id) {
         return repository.findById(id).orElseThrow();
     }
 
-    public TypeEmploye updateTypeEmploye(Long id, TypeEmploye typeEmploye) {
+    public TypeEmploye updateTypeEmploye(int id, TypeEmploye typeEmploye) {
         TypeEmploye existingTypeEmploye = repository.findById(id).orElseThrow();
         existingTypeEmploye.setDesignation(typeEmploye.getDesignation());
         existingTypeEmploye.setSalaireBase(typeEmploye.getSalaireBase());
@@ -34,7 +34,7 @@ public class TypeEmployeService {
         return repository.save(existingTypeEmploye);
     }
 
-    public void deleteTypeEmploye(Long id) {
+    public void deleteTypeEmploye(int id) {
         repository.deleteById(id);
     }
 }
